+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - 
 
 #### New Features & Functionality
+- Add plugin component.
 
 - QueryTemplate component